+++ src/lib/components/ImageViewer.svelte
@@ -21,6 +21,7 @@
 <!-- svelte-ignore a11y_no_noninteractive_element_interactions -->
 <div
   class="image-viewer"
+  data-region="canvas"
   bind:this={frame}
   tabindex="-1"
   role="dialog"